Integrator: When activating the `stokes` and `moment` integrators, make sure that `stokes` comes last
This commit works around a side effect of mitsuba-renderer/mitsuba3#1668.
That change makes it impossible to nest the `stokes` integrator under
another one. This, in practice, requires to make `stokes` the top-level
integrator. Therefore, `moment` can only be applied to variables output
by the main integrator, i.e. the I component of the Stokes vector.

A consequence is that variance can no longer be computed for the Q, U
and V components. This introduces further complications in the
post-processing pipeline, which used to assume that the m2 and mean
variables shared the same dimensions: this no longer applies, requiring
additional workarounds in post-processing logic. The proposed patch is
not ideal, but a more robust solution would require further refactoring
of the pipeline.