Phase function parameters are no longer nested
This commit changes the path under which atmospheric phase function
parameters appear in the parameter table: the phase function now always
appears at the top level, at least when declared prior to the medium.

Note that the parameter search strategy is still here: this is a
conservative fix. Judging by the tests that have been carried out
outside of Eradiate, we might even consider to remove all parameter
search protocols in the future.